requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
ffmpeg-python==0.2.0
filelock==3.11.0
future==0.18.3
idna==3.4
Jinja2==3.1.2
llvmlite==0.39.1
MarkupSafe==2.1.2
more-itertools==9.1.0
mpmath==1.3.0
networkx==3.1
numba==0.56.4
numpy==1.23.5
openai-whisper==20230314
regex==2023.3.23
requests==2.28.2
sympy==1.11.1
tiktoken==0.3.1
torch==2.0.0
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
